backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.2",
    "langchain-groq>=0.2.0",
    "langchain-mongodb>=0.2.0",
    "langchain>=0.3.3",
    "pymongo>=4.10.1",
    "python-dotenv>=1.0.1",
]
